requirements.txt
filelock==3.15.4
fsspec==2024.6.1
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
sympy==1.13.1
torch==2.4.0
triton==3.0.0
typing-extensions==4.12.2
tensortrade
tensorflow
stable-baselines3
ccxt
flask
python-binance
